refactor: split verify package
