+++ internal/server/register_routes.go
@@ -8,17 +8,19 @@ import (
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/message"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/object"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/packages"
-	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/photographer"
+	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/photographers"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/quotation"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/review"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/user"
+	"github.com/CP-RektMart/pic-me-pls-backend/internal/services/verify"
 )
 
 func (s *Server) RegisterRoutes(
 	authMiddleware authentication.AuthMiddleware,
 	authHandler *auth.Handler,
 	userHandler *user.Handler,
-	photographerHandler *photographer.Handler,
+	photographersHandler *photographers.Handler,
+	verifyHandler *verify.Handler,
 	packagesHandler *packages.Handler,
 	reviewHandler *review.Handler,
 	categoryHandler *category.Handler,
@@ -27,29 +29,28 @@ func (s *Server) RegisterRoutes(
 	quotationHandler *quotation.Handler,
 	mediaHandler *media.Handler,
 ) {
-	api := s.app.Group("/api")
-	v1 := api.Group("/v1")
+	v1 := s.app.Group("/api/v1")
 
-	// auth
+	// all, auth
 	auth := v1.Group("/auth")
 	auth.Post("/login", authHandler.HandleLogin)
 	auth.Post("/register", authHandler.HandleRegister)
 	auth.Post("/refresh-token", authHandler.HandleRefreshToken)
 	auth.Post("/logout", authMiddleware.Auth, authHandler.HandleLogout)
 
-	// user
+	// all, user
 	v1.Get("/me", authMiddleware.Auth, userHandler.HandleGetMe)
 	v1.Patch("/me", authMiddleware.Auth, userHandler.HandleUpdateMe)
 
 	// verify citizen card
 	photographer := v1.Group("/photographer")
-	photographer.Get("/citizen-card", authMiddleware.AuthPhotographer, photographerHandler.HandleGetCitizenCard)
-	photographer.Post("/verify", authMiddleware.AuthPhotographer, photographerHandler.HandleVerifyCard)
-	photographer.Patch("/reverify", authMiddleware.AuthPhotographer, photographerHandler.HandleReVerifyCard)
+	photographer.Get("/citizen-card", authMiddleware.AuthPhotographer, verifyHandler.HandleGetCitizenCard)
+	photographer.Post("/verify", authMiddleware.AuthPhotographer, verifyHandler.HandleVerifyCard)
+	photographer.Patch("/reverify", authMiddleware.AuthPhotographer, verifyHandler.HandleReVerifyCard)
 
 	// get photographer
 	photographers := v1.Group("/photographers")
-	photographers.Get("/", photographerHandler.HandleGetAllPhotographers)
+	photographers.Get("/", photographersHandler.HandleGetAllPhotographers)
 
 	auth.Post("/logout", authMiddleware.Auth, authHandler.HandleLogout)
