+++ internal/services/photographers/get-photographer.go
@@ -1,4 +1,4 @@
-package photographer
+package photographers
 
 import (
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/dto"
