+++ internal/services/verify/reverify-card.go
@@ -1,4 +1,4 @@
-package photographer
+package verify
 
 import (
 	"mime/multipart"
