+++ internal/services/verify/verify-card.go
@@ -1,4 +1,4 @@
-package photographer
+package verify
 
 import (
 	"context"
